refs rate of interest datatype changes commit

+++ src/main/java/shfl/st/lap/disbursementrequest/model/DisbursementModel.java
@@ -13,7 +13,7 @@ public class DisbursementModel {
 	private float earlierDisbAmt;
 	private float disbAmt;
 	private int disbNo;
-	private int rateOfInterest;
+	private float rateOfInterest;
 	private float totalDisbAmt;
 	private Date dateOfDisb;
 	private String billingDay;
